Algae Collection Skeleton Changes 1
Made commands to set the motors to null. Renamed and reorganized some packages and commands. Added more information to the motor.

No questions.

No tests.
