+++ src/main/java/competition/electrical_contract/ElectricalContract.java
@@ -3,6 +3,7 @@ package competition.electrical_contract;
 import xbot.common.injection.electrical_contract.CANMotorControllerInfo;
 import xbot.common.injection.electrical_contract.CANTalonInfo;
 import xbot.common.injection.electrical_contract.DeviceInfo;
+import xbot.common.injection.electrical_contract.MotorControllerType;
 import xbot.common.injection.electrical_contract.XCameraElectricalContract;
 import xbot.common.injection.electrical_contract.XSwerveDriveElectricalContract;
 import xbot.common.injection.swerve.SwerveInstance;
@@ -16,8 +17,6 @@ public abstract class ElectricalContract implements XSwerveDriveElectricalContra
     public abstract CANMotorControllerInfo getDriveMotor(SwerveInstance swerveInstance);
 
     public abstract CANMotorControllerInfo getSteeringMotor(SwerveInstance swerveInstance);
-
-    public abstract CANMotorControllerInfo getAlgaeCollectionMotor();
 
     public abstract DeviceInfo getSteeringEncoder(SwerveInstance swerveInstance);
