+++ src/main/java/competition/subsystems/algae_collection/commands/AlgaeCollectionStopCommand.java
@@ -1,0 +1,21 @@
+package competition.subsystems.algae_collection.commands;
+
+import competition.subsystems.algae_collection.AlgaeCollectionSubsystem;
+import xbot.common.command.BaseCommand;
+
+import javax.inject.Inject;
+
+public class AlgaeCollectionStopCommand extends BaseCommand {
+    AlgaeCollectionSubsystem algae;
+
+    @Inject
+    public AlgaeCollectionStopCommand(AlgaeCollectionSubsystem algaeCollectionSubsystem) {
+        algae = algaeCollectionSubsystem;
+        this.addRequirements(algae);
+    }
+
+    @Override
+    public void initialize() {
+        algae.stop();
+    }
+}
